+++ RHRouter.cpp
@@ -17,7 +17,7 @@ RHRouter::RoutedMessage RHRouter::_tmpMessage;
 
 ////////////////////////////////////////////////////////////////////
 // Constructors
-RHRouter::RHRouter(RHGenericDriver& driver, uint8_t thisAddress) 
+RHRouter::RHRouter(RHGenericDriver& driver, uint32_t thisAddress) 
     : RHReliableDatagram(driver, thisAddress)
 {
     _max_hops = RH_DEFAULT_MAX_HOPS;
@@ -47,7 +47,7 @@ void RHRouter::setIsaRouter(bool isa_router)
     _isa_router = isa_router;
 }
 ////////////////////////////////////////////////////////////////////
-void RHRouter::addRouteTo(uint8_t dest, uint8_t next_hop, uint8_t state)
+void RHRouter::addRouteTo(uint32_t dest, uint32_t next_hop, uint8_t state)
 {
     uint8_t i;
 
@@ -90,7 +90,7 @@ void RHRouter::addRouteTo(uint8_t dest, uint8_t next_hop, uint8_t state)
 }
 
 ////////////////////////////////////////////////////////////////////
-RHRouter::RoutingTableEntry* RHRouter::getRouteTo(uint8_t dest)
+RHRouter::RoutingTableEntry* RHRouter::getRouteTo(uint32_t dest)
 {
     uint8_t i;
     for (i = 0; i < RH_ROUTING_TABLE_SIZE; i++)
@@ -127,7 +127,7 @@ void RHRouter::printRoutingTable()
 }
 
 ////////////////////////////////////////////////////////////////////
-bool RHRouter::deleteRouteTo(uint8_t dest)
+bool RHRouter::deleteRouteTo(uint32_t dest)
 {
     uint8_t i;
     for (i = 0; i < RH_ROUTING_TABLE_SIZE; i++)
@@ -157,14 +157,14 @@ void RHRouter::clearRoutingTable()
 }
 
 
-uint8_t RHRouter::sendtoWait(uint8_t* buf, uint8_t len, uint8_t dest, uint8_t flags)
+uint8_t RHRouter::sendtoWait(uint8_t* buf, uint8_t len, uint32_t dest, uint8_t flags)
 {
     return sendtoFromSourceWait(buf, len, dest, _thisAddress, flags);
 }
 
 ////////////////////////////////////////////////////////////////////
 // Waits for delivery to the next hop (but not for delivery to the final destination)
-uint8_t RHRouter::sendtoFromSourceWait(uint8_t* buf, uint8_t len, uint8_t dest, uint8_t source, uint8_t flags)
+uint8_t RHRouter::sendtoFromSourceWait(uint8_t* buf, uint8_t len, uint32_t dest, uint32_t source, uint8_t flags)
 {
     if (((uint16_t)len + sizeof(RoutedMessageHeader)) > _driver.maxMessageLength())
 	return RH_ROUTER_ERROR_INVALID_LENGTH;
@@ -184,7 +184,7 @@ uint8_t RHRouter::sendtoFromSourceWait(uint8_t* buf, uint8_t len, uint8_t dest, 
 uint8_t RHRouter::route(RoutedMessage* message, uint8_t messageLen)
 {
     // Reliably deliver it if possible. See if we have a route:
-    uint8_t next_hop = RH_BROADCAST_ADDRESS;
+    uint32_t next_hop = RH_BROADCAST_ADDRESS;
     if (message->header.dest != RH_BROADCAST_ADDRESS)
     {
 	RoutingTableEntry* route = getRouteTo(message->header.dest);
@@ -209,11 +209,11 @@ void RHRouter::peekAtMessage(RoutedMessage* message, uint8_t messageLen)
 }
 
 ////////////////////////////////////////////////////////////////////
-bool RHRouter::recvfromAck(uint8_t* buf, uint8_t* len, uint8_t* source, uint8_t* dest, uint8_t* id, uint8_t* flags)
+bool RHRouter::recvfromAck(uint8_t* buf, uint8_t* len, uint32_t* source, uint32_t* dest, uint8_t* id, uint8_t* flags)
 {  
-    uint8_t tmpMessageLen = sizeof(_tmpMessage);
-    uint8_t _from;
-    uint8_t _to;
+    uint16_t tmpMessageLen = sizeof(_tmpMessage);
+    uint32_t _from;
+    uint32_t _to;
     uint8_t _id;
     uint8_t _flags;
     if (RHReliableDatagram::recvfromAck((uint8_t*)&_tmpMessage, &tmpMessageLen, &_from, &_to, &_id, &_flags))
@@ -299,7 +299,7 @@ bool RHRouter::recvfromAck(uint8_t* buf, uint8_t* len, uint8_t* source, uint8_t*
 }
 
 ////////////////////////////////////////////////////////////////////
-bool RHRouter::recvfromAckTimeout(uint8_t* buf, uint8_t* len, uint16_t timeout, uint8_t* source, uint8_t* dest, uint8_t* id, uint8_t* flags)
+bool RHRouter::recvfromAckTimeout(uint8_t* buf, uint8_t* len, uint16_t timeout, uint32_t* source, uint32_t* dest, uint8_t* id, uint8_t* flags)
 {  
     unsigned long starttime = millis();
     int32_t timeLeft;
