+++ RH_CC110.cpp
@@ -283,7 +283,7 @@ void RH_CC110::clearRxBuf()
     ATOMIC_BLOCK_END;
 }
 
-bool RH_CC110::recv(uint8_t* buf, uint8_t* len)
+bool RH_CC110::recv(uint8_t* buf, uint16_t* len)
 {
     if (!available())
 	return false;
