+++ RH_E32.cpp
@@ -239,7 +239,7 @@ bool RH_E32::available()
     return _rxBufValid;
 }
 
-bool RH_E32::recv(uint8_t* buf, uint8_t* len)
+bool RH_E32::recv(uint8_t* buf, uint16_t* len)
 {
     if (!available())
 	return false;
